type:change(merci seigneur)

+++ database/migrations/2024_02_28_135948_create_analyse_data_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('sexe');
             $table->string('age');
             $table->string('type');
-            $table->integer('taille')->nullable();
+            $table->float('taille')->nullable();
             $table->integer('poids')->nullable();
             $table->string('systolique')->nullable();
             $table->string('diastolique')->nullable();
